+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - scripts
 - ability to view and edit user data
 - theme preview in theme change section
+- ability to change user password